Cherry Pick changes to 2.1 (#414)
* [SPDM] Add support for END_SESSION command (#350)

* secure message encrypt decrypt and authentication support

---------

Co-authored-by: Parvathi Bhogaraju <pbhogaraju@microsoft.com>

* [fpga] remove scripts for dispatching FPGA tests (#379)

These shell scripts have been replaced by an xtask flow in #355, #357,
and #361.

Signed-off-by: Tim Trippel <ttrippel@google.com>

* [xtask] enable dispatching a single FPGA test (#380)

This updates the xtask flow for executing FPGA tests by enabling running
a single specified task instead of the entire test suite.

Signed-off-by: Tim Trippel <ttrippel@google.com>

* [xtask] enable streaming test output (#382)

Add an option the FPGA test xtask flow to stream test output to the
console for test debugging.

Signed-off-by: Tim Trippel <ttrippel@google.com>

* Add ROM feature to enable easier Core testing (#388)

This feature loops forever in ROM once the core fuses have been set and core has been released from reset.

chipsalliance/caliptra-sw#2391 has an example of passing tests with this ROM.

* Pass SoC Manifest SVN fuse values to Caliptra (#377)

* Pass SoC Manifest SVN fuse values to Caliptra

- MCU ROM retrieves SoC Manifest SVN and Max SVN values and passes it to
Caliptra using SoC interface.
- Add logic in MCU ROM to monitor Caliptra for Fatal Errors while
waiting for FW to be ready. Caliptra will signal a fatal error if the
SVN check fails. In this way, MCU will propagate the fatal error to SoC.
- Added SoC Manifest SVN to the OTP emulator peripheral
- Add arguments to the emulator runner to pass SoC Manifest SVN
- Add integration test to verify SoC Manifest Verification

* Fix failing FPGA reg_access unit test

Need to specify explicitly pqc_type and also speficy the PK values in
the boot params for fuses

* Allow Core Test ROM to continue when a GPIO is asserted. (#389)

MCU ROM needs to load mutable firmware for many of the tests. This will allow the ROM used for testing Caliptra Core to continue if the test needs mutable firmware.

* [SPDM] Enable Secure session tests (#378)

* secure message encrypt decrypt and authentication support
---------

Co-authored-by: Parvathi Bhogaraju <pbhogaraju@microsoft.com>

* MCTP-related I3C fixes (#387)

* Pre-emptively send I3C Private Read directly to the i3c-core registers
  (which queues it for sending) rather than waiting for an interrupt. In
  reality, we'll never be able to respond to the interrupt in time to
  fill the buffer otherwise, since the buffer needs to be ready within
  just a few clock cycles.
* Use the latest caliptra-sw hardware model with new I3C methods for
  testing https://github.com/chipsalliance/caliptra-sw/pull/2424
* MCTP GetVersion version fields should be in big endian, not little, to
  match the spec
* Emulator I3C peripheral should honor the interrupt enable register
* I3C interrupts need to actually enabled via the MCTP mux in the
  runtime board file
* Add MCTP GetVersion test for FPGA
* Remove unused I3C controller code now that it is in caliptra-sw
* Fix MCI mailbox address to use platform-specific address instead of
  emulator
* Check response and request in MCTP capsule
* When waiting for IBI, loop until we stop receiving packets so
  that we don't have to rely on retries if there are packets ready
  to be read. (This makes some PLDM tests less flaky.)
* Lock use of the alarm capsule in userspace to prevent tasks from
  overwriting each others
* Increase I3C alarm amount so as not to accidentally starve the user
  process if the kernel is too busy handling short timers
* Fix a few edge cases around buffering of messages in the capsule
* Temporarily disable FPGA tests

* [fpga] fix JTAG pin orderings (#397)

This aligns the JTAG pin orderings across all TAPs (core, lcc, and mcu) as:
1. tck
2. tms
3. tdi
4. trst
5. tdo

Additionally, this fixes the openocd configuration script to use the
same JTAG pin orderings. This partially addresses #363 by enabling
connecting to the LCC TAP (along with the other TAPs).

With this fix, I am able to successfully connect to the LCC TAP with
OpenOCD.

Signed-off-by: Tim Trippel <ttrippel@google.com>

* Rsync FPGA files to the local directory (#404)

* [xtask]: Optionally pass feature flags to rom and runtime builder for all build … (#407)

* [xtask]: Pass feature flags to rom and runtime builder for all build command

* Update builder/src/all.rs

Co-authored-by: Zach Halvorsen <zhalvorsen@google.com>

* Add caliptra-sw argument to `xtask fpga built-test`. (#408)

This optional argument makes it easier to test local caliptra-sw changes on the FPGA using xtask.

* Xtask: Add support for running "core" tests on a subsystem FPGA (#409)

* Refactor Vendor Defined message handlers (#391)

* refactor vdm handlers

---------

Co-authored-by: Parvathi Bhogaraju <pbhogaraju@microsoft.com>

* Set AXI user to Core test default. (#401)

Several core tests assume the AXI user is `0x1`. This sets the default in MCU FPGA user to that value. This makes several more tests pass.

At some point this should be communicated to the MCU ROM in a configurable way.

* Temporarily change AXI user in caliptra-sw to pass CI

* [Core test ROM] pause after Caliptra boot go (#410)

This adds another Core test MCU ROM pause point. This will allow us
to test flows that just remove the SoC from reset without booting the
full system. The new flow is:

1. Come out of reset
2. Wait for signal from Core
3. Program fuses and initialize the other peripherals
4. Wait for signal from Core
5. Load mutable firmware

* Add MCU SVN verification (#390)

* Add MCU SVN verification

Provide an example implementation of how SVN can be added to an MCU
image

- Add a header that contains an SVN to MCU FW image. This package
(header and FW binary) will be loaded in the MCU SRAM. MCU ROM will
verify the SVN in the image header and jump to the offset where the FW
binary is found.
- Use Vendor Hash Prod fuses to implement MCU SVNs
- Verify MCU SVN in ROM, user must implement the ImageVerifier trait
that implements the SVN verification. An example implementation is
provided in mcu_imager_verifier.rs

* Correct typo

* Fix FPGA HW Model being unable to run recovery flow (#406)

The default lifecycle state was being set to prod which would later cause it to fail the Caliptra firmware integrity checks because the built image uses ML-DSA by default, but the hardware model uses LMS by default.

* Update caliptra-sw dependency

* Address caliptra-sw build errors

---------

Signed-off-by: Tim Trippel <ttrippel@google.com>
Co-authored-by: Parvathi Bhogaraju <bhogaraju.parvathi@gmail.com>
Co-authored-by: Parvathi Bhogaraju <pbhogaraju@microsoft.com>
Co-authored-by: Timothy Trippel <5633066+timothytrippel@users.noreply.github.com>
Co-authored-by: Zach Halvorsen <zhalvorsen@google.com>
Co-authored-by: mlvisaya <38512415+mlvisaya@users.noreply.github.com>
Co-authored-by: Christopher Swenson <chswenson@microsoft.com>