Cherry pick main changes into main-2.1 (#444)
* Enable Warm Reset (#298)

* Implement MCU warm reset via RESET_REQUEST register

- Add RESET_REQUEST register support in MCI peripheral
- Update RESET_REASON when reset is requested
- Check for reset request in emulator step loop
- Trigger CPU reset when mcu_req bit is set
- Add registers-generated dependency for field access

* rom: add warm boot flow

Add warm boot module, and invoke it from rom. Warm boot doesn't do much
else other than sanity check that runtime firmware is loaded and jumps
to firmware.
Also added a helper function mci to trigger a warm reset by writing to
the reset request register.

* fix: ResetReason emulator

Remove use of first_mci_reset_captured. The flag was unnecessary and
caused the need to write to reset request register twice to change the
value. Also fixed tests to account for change.

* test: warm boot flow

Add simple test to trigger warm boot after exiting ROM and entering
kernel.

* test: update warm reset test to use mcu capsule components

* Clear Caliptra mailbox lock on warm boot

* Fix review comments.

- move use statements.
- remove conflict markers from rebase.
- jump past header during warm boot.

* Add ECDSA verification for signature checks in tests. (#441)

Co-authored-by: Parvathi Bhogaraju <pbhogaraju@microsoft.com>

* Onboard mctp-capsule-loopback test to FPGA CI (#437)

The `FwId` and `REGISTERED_FW` abstraction doesn't quite work for specifying the different runtime binaries since they are in a different crate for FPGA. There's additional complexity in that a different runtime would also require a different SoC manifest to authorize it. So I added a slightly different mechanism for building and packaging them in the the `all-fw.zip` bundle.

The downside is that the FPGA test is flaky due to it halting when going through the full recovery flow to load MCU runtime.

* emulator: Support UTF8 characters in output (#442)

So that the Tock panic prints correctly finally:

```
╔═══════════╤══════════════════════════════════════════╗
║  Address  │ Region Name    Used | Allocated (bytes)  ║
╚0x4007F4B0═╪══════════════════════════════════════════╝
            │ Grant Ptrs      120
            │ Upcalls         320
            │ Process         756
 0x4007F004 ┼───────────────────────────────────────────
```

I'll also port this to `caliptra-sw` for the HwModel output.

* Bump caliptra-sw commit

---------

Co-authored-by: Raghu Krishnamurthy <13037141+raghuncstate@users.noreply.github.com>
Co-authored-by: Parvathi Bhogaraju <bhogaraju.parvathi@gmail.com>
Co-authored-by: Parvathi Bhogaraju <pbhogaraju@microsoft.com>
Co-authored-by: Christopher Swenson <chswenson@microsoft.com>